Disable history sharing between instances of zsh
Thought it would be useful. Turns out, I split my terminals
and start different instances of zsh in order to work on
*different* things. So I don't want their history to be merged
as long as they are split. And zsh is smart enough to only do the
merge when both shells exit, whilst still saving to disk every
time a command is started.